rag-server/requirements.txt
# LangChain 핵심
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-upstage>=0.0.1
langchain-chroma>=0.1.0
langchain-community>=0.0.20

# 유틸리티
python-dotenv>=1.0.0
pydantic>=2.0.0
boto3>=1.28.0

# 문서 처리 (전처리 시에만 필요)
pypdf>=3.0.0
langchain-text-splitters>=0.0.1

# fastapi 추가 의존성 설치
fastapi[standard]